+++ src/main/java/com/direwolf20/buildinggadgets/common/items/gadgets/AbstractGadget.java
@@ -24,6 +24,7 @@ import com.direwolf20.buildinggadgets.common.util.lang.MessageTranslation;
 import com.direwolf20.buildinggadgets.common.util.lang.Styles;
 import com.direwolf20.buildinggadgets.common.util.lang.TooltipTranslation;
 import com.direwolf20.buildinggadgets.common.util.ref.NBTKeys;
+import net.minecraft.block.Block;
 import com.direwolf20.buildinggadgets.common.util.tools.CapabilityUtil.EnergyUtil;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedSet;
@@ -33,6 +34,9 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.tags.BlockTags.Wrapper;
+import net.minecraft.tags.Tag;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.util.math.ChunkPos;
@@ -59,18 +63,29 @@ import java.util.function.Supplier;
 import static com.direwolf20.buildinggadgets.common.util.GadgetUtils.withSuffix;
 
 public abstract class AbstractGadget extends Item {
-
     private BaseRenderer renderer;
+    private final Tag<Block> whiteList;
+    private final Tag<Block> blackList;
     private Supplier<UndoWorldSave> saveSupplier;
 
-    public AbstractGadget(Properties builder, IntSupplier undoLengthSupplier, String undoName) {
+    public AbstractGadget(Properties builder, IntSupplier undoLengthSupplier, String undoName, ResourceLocation whiteListTag, ResourceLocation blackListTag) {
         super(builder);
         renderer = DistExecutor.runForDist(this::createRenderFactory, () -> () -> null);
+        this.whiteList = new Wrapper(whiteListTag);
+        this.blackList = new Wrapper(blackListTag);
         saveSupplier = SaveManager.INSTANCE.registerUndoSave(w -> SaveManager.getUndoSave(w, undoLengthSupplier, undoName));
     }
 
     public abstract int getEnergyMax();
     public abstract int getEnergyCost(ItemStack tool);
+
+    public Tag<Block> getWhiteList() {
+        return whiteList;
+    }
+
+    public Tag<Block> getBlackList() {
+        return blackList;
+    }
 
     @OnlyIn(Dist.CLIENT)
     public BaseRenderer getRender() {
@@ -140,6 +155,20 @@ public abstract class AbstractGadget extends Item {
     @Override
     public boolean getIsRepairable(ItemStack toRepair, ItemStack repair) {
         return !EnergyUtil.hasCap(toRepair) && repair.getItem() == Items.DIAMOND;
+    }
+
+    public boolean isAllowedBlock(Block block) {
+        if (getWhiteList().getAllElements().isEmpty())
+            return ! getBlackList().contains(block);
+        return getWhiteList().contains(block);
+    }
+
+    public boolean isAllowedBlock(Block block, @Nullable PlayerEntity notifiedPlayer) {
+        if (isAllowedBlock(block))
+            return true;
+        if (notifiedPlayer != null)
+            notifiedPlayer.sendStatusMessage(new StringTextComponent(TextFormatting.RED + new TranslationTextComponent("message.gadget.invalidblock").getUnformattedComponentText()), true);
+        return false;
     }
 
     public static ItemStack getGadget(PlayerEntity player) {