+++ src/main/java/com/direwolf20/buildinggadgets/common/items/gadgets/GadgetExchanger.java
@@ -24,6 +24,8 @@ import com.direwolf20.buildinggadgets.common.util.lang.LangUtil;
 import com.direwolf20.buildinggadgets.common.util.lang.MessageTranslation;
 import com.direwolf20.buildinggadgets.common.util.lang.Styles;
 import com.direwolf20.buildinggadgets.common.util.lang.TooltipTranslation;
+import com.direwolf20.buildinggadgets.common.util.ref.Reference.BlockReference.TagReference;
+import com.direwolf20.buildinggadgets.common.util.tools.modes.ExchangingMode;
 import com.direwolf20.buildinggadgets.common.world.FakeBuilderWorld;
 import com.google.common.collect.ImmutableMultiset;
 import net.minecraft.block.BlockState;
@@ -68,7 +70,7 @@ public class GadgetExchanger extends ModeGadget {
     private static final FakeBuilderWorld fakeWorld = new FakeBuilderWorld();
 
     public GadgetExchanger(Properties builder, IntSupplier undoLengthSupplier, String undoName) {
-        super(builder, undoLengthSupplier, undoName);
+        super(builder, undoLengthSupplier, undoName, TagReference.WHITELIST_EXCHANGING, TagReference.BLACKLIST_EXCHANGING);
     }
 
     @Override